refactor: move i/n/quicx into i/model
We want to have a model containing all the shared data structures
and interfaces that make OONI.

We start from moving internal/netxlite/quicx into this new package
and then we'll follow suit by moving other pieces of code.
