+++ internal/engine/legacy/errorsx/quic_test.go
@@ -9,15 +9,15 @@ import (
 	"testing"
 
 	"github.com/lucas-clemente/quic-go"
+	"github.com/ooni/probe-cli/v3/internal/model"
 	"github.com/ooni/probe-cli/v3/internal/netxlite"
 	"github.com/ooni/probe-cli/v3/internal/netxlite/mocks"
-	"github.com/ooni/probe-cli/v3/internal/netxlite/quicx"
 )
 
 func TestErrorWrapperQUICListenerSuccess(t *testing.T) {
 	ql := &ErrorWrapperQUICListener{
 		QUICListener: &mocks.QUICListener{
-			MockListen: func(addr *net.UDPAddr) (quicx.UDPLikeConn, error) {
+			MockListen: func(addr *net.UDPAddr) (model.UDPLikeConn, error) {
 				return &net.UDPConn{}, nil
 			},
 		},
@@ -32,7 +32,7 @@ func TestErrorWrapperQUICListenerSuccess(t *testing.T) {
 func TestErrorWrapperQUICListenerFailure(t *testing.T) {
 	ql := &ErrorWrapperQUICListener{
 		QUICListener: &mocks.QUICListener{
-			MockListen: func(addr *net.UDPAddr) (quicx.UDPLikeConn, error) {
+			MockListen: func(addr *net.UDPAddr) (model.UDPLikeConn, error) {
 				return nil, io.EOF
 			},
 		},
