+++ internal/measurex/quic.go
@@ -13,12 +13,12 @@ import (
 	"time"
 
 	"github.com/lucas-clemente/quic-go"
+	"github.com/ooni/probe-cli/v3/internal/model"
 	"github.com/ooni/probe-cli/v3/internal/netxlite"
-	"github.com/ooni/probe-cli/v3/internal/netxlite/quicx"
 )
 
 // QUICConn is the kind of conn used by QUIC.
-type QUICConn = quicx.UDPLikeConn
+type QUICConn = model.UDPLikeConn
 
 // QUICDialer creates QUICSesssions.
 type QUICDialer = netxlite.QUICDialer
@@ -45,7 +45,7 @@ func (ql *quicListenerDB) Listen(addr *net.UDPAddr) (QUICConn, error) {
 }
 
 type udpLikeConnDB struct {
-	quicx.UDPLikeConn
+	model.UDPLikeConn
 	begin time.Time
 	db    WritableDB
 }
