+++ internal/model/doc.go
@@ -1,0 +1,2 @@
+// Package model contains the shared interfaces and data structures.
+package model
